requirements.txt
Django
Pillow
requests
python-dateutil
psycopg2-binary
gunicorn
django-cockroachdb==3.1.*
